backend_py/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
python-dotenv>=1.0.1
matplotlib>=3.6.3
jupyterlab>=3.5.3
music21>=8.1.0
ipywidgets>=8.0.4
tensorboard>=2.10.1
torchvision==0.16.2
torchaudio==2.1.2
diffusers[torch]
transformers[torch]
accelerate[torch]
scipy
safetensors[torch]